+++ src/freenet/node/fcp/UploadFileRequestStatus.java
@@ -20,7 +20,7 @@ public class UploadFileRequestStatus extends UploadRequestStatus {
 			boolean success, int total, int min, int fetched, int fatal, int failed,
 			boolean totalFinalized, long last, short prio, // all these passed to parent
 			FreenetURI finalURI, FreenetURI targetURI, 
-			short failureCode, String failureReasonShort, String failureReasonLong,
+			int failureCode, String failureReasonShort, String failureReasonLong,
 			long dataSize, String mimeType, File origFilename, COMPRESS_STATE compressing) {
 		super(identifier, persistence, started, finished, success, total, min, fetched, 
 				fatal, failed, totalFinalized, last, prio, finalURI, targetURI,
